LayoutTests/http/wpt/loading/early-hints/preconnect-csp-blocked.h2.window-expected.txt

CONSOLE MESSAGE: Refused to connect to https://not-web-platform.test:9000/ because it appears in neither the connect-src directive nor the default-src directive of the Content Security Policy.


PASS Early hints preconnect
